+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/asn1/Asn1BitString.kt
@@ -82,7 +82,7 @@ class Asn1BitString private constructor(
 
         @Throws(Asn1Exception::class)
         private fun decode(src: Asn1Primitive, tagOverride: TLV.Tag? = null): Asn1BitString {
-            val expected = tagOverride ?: TLV.Tag(BERTags.BIT_STRING.toUInt(), constructed = false)
+            val expected = tagOverride ?: TLV.Tag(BERTags.BIT_STRING.toULong(), constructed = false)
             if (src.tag != expected)
                 throw Asn1TagMismatchException(expected, src.tag)
             if (src.length == 0) return Asn1BitString(0, byteArrayOf())
@@ -97,7 +97,7 @@ class Asn1BitString private constructor(
         override fun decodeFromTlv(src: Asn1Primitive, tagOverride: TLV.Tag?) = decode(src, tagOverride)
     }
 
-    override fun encodeToTlv() = Asn1Primitive(BERTags.BIT_STRING.toUInt(), byteArrayOf(numPaddingBits, *rawBytes))
+    override fun encodeToTlv() = Asn1Primitive(BERTags.BIT_STRING.toULong(), byteArrayOf(numPaddingBits, *rawBytes))
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
         if (other == null || this::class != other::class) return false
