+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/asn1/Asn1Encoding.kt
@@ -191,7 +191,7 @@ object Asn1 {
      * }
      *  ```
      */
-    fun Tagged(tag: UInt, root: Asn1TreeBuilder.() -> Unit): Asn1Tagged {
+    fun Tagged(tag: ULong, root: Asn1TreeBuilder.() -> Unit): Asn1Tagged {
         val seq = Asn1TreeBuilder()
         seq.root()
         return Asn1Tagged(tag, seq.elements)
@@ -200,13 +200,13 @@ object Asn1 {
     /**
      * Exception-free version of [Tagged]
      */
-    fun TaggedOrNull(tag: UInt, root: Asn1TreeBuilder.() -> Unit) =
+    fun TaggedOrNull(tag: ULong, root: Asn1TreeBuilder.() -> Unit) =
         catching { Tagged(tag, root) }.getOrNull()
 
     /**
      * Safe version on [Tagged], wrapping the result into a [KmmResult]
      */
-    fun TaggedSafe(tag: UInt, root: Asn1TreeBuilder.() -> Unit) =
+    fun TaggedSafe(tag: ULong, root: Asn1TreeBuilder.() -> Unit) =
         catching { Tagged(tag, root) }
 
 
@@ -264,7 +264,7 @@ object Asn1 {
     /**
      * Adds a NULL [Asn1Primitive] to this ASN.1 structure
      */
-    fun Null() = Asn1Primitive(ASN1_NULL.toUInt(), byteArrayOf())
+    fun Null() = Asn1Primitive(ASN1_NULL.toULong(), byteArrayOf())
 
 
     /**
@@ -304,23 +304,23 @@ object Asn1 {
 /**
  * Produces an INTEGER as [Asn1Primitive]
  */
-fun Int.encodeToTlv() = Asn1Primitive(INTEGER.toUInt(), encodeToDer())
+fun Int.encodeToTlv() = Asn1Primitive(INTEGER.toULong(), encodeToDer())
 
 
 /**
  * Produces a BOOLEAN as [Asn1Primitive]
  */
-fun Boolean.encodeToTlv() = Asn1Primitive(BOOLEAN.toUInt(), byteArrayOf(if (this) 0xff.toByte() else 0))
+fun Boolean.encodeToTlv() = Asn1Primitive(BOOLEAN.toULong(), byteArrayOf(if (this) 0xff.toByte() else 0))
 
 /**
  * Produces an INTEGER as [Asn1Primitive]
  */
-fun Long.encodeToTlv() = Asn1Primitive(INTEGER.toUInt(), encodeToDer())
+fun Long.encodeToTlv() = Asn1Primitive(INTEGER.toULong(), encodeToDer())
 
 /**
  * Produces an INTEGER as [Asn1Primitive]
  */
-fun BigInteger.encodeToTlv() = Asn1Primitive(INTEGER.toUInt(), toTwosComplementByteArray())
+fun BigInteger.encodeToTlv() = Asn1Primitive(INTEGER.toULong(), toTwosComplementByteArray())
 
 /**
  * Produces an OCTET STRING as [Asn1Primitive]
@@ -330,7 +330,7 @@ fun ByteArray.encodeToTlvOctetString() = Asn1PrimitiveOctetString(this)
 /**
  * Produces a BIT STRING as [Asn1Primitive]
  */
-fun ByteArray.encodeToTlvBitString() = Asn1Primitive(BIT_STRING.toUInt(), encodeToBitString())
+fun ByteArray.encodeToTlvBitString() = Asn1Primitive(BIT_STRING.toULong(), encodeToBitString())
 
 /**
  * Prepends 0x00 to this ByteArray for encoding it into a BIT STRING. Useful for implicit tagging
@@ -347,13 +347,13 @@ private fun Long.encodeToDer() = if (this == 0L) byteArrayOf(0) else
  * Produces a UTC TIME as [Asn1Primitive]
  */
 fun Instant.encodeToAsn1UtcTime() =
-    Asn1Primitive(UTC_TIME.toUInt(), encodeToAsn1Time().drop(2).encodeToByteArray())
+    Asn1Primitive(UTC_TIME.toULong(), encodeToAsn1Time().drop(2).encodeToByteArray())
 
 /**
  * Produces a GENERALIZED TIME as [Asn1Primitive]
  */
 fun Instant.encodeToAsn1GeneralizedTime() =
-    Asn1Primitive(GENERALIZED_TIME.toUInt(), encodeToAsn1Time().encodeToByteArray())
+    Asn1Primitive(GENERALIZED_TIME.toULong(), encodeToAsn1Time().encodeToByteArray())
 
 private fun Instant.encodeToAsn1Time(): String {
     val value = this.toString()
